requirements.txt
PyYAML>=6.0
requests>=2.28.0
Jinja2>=3.0.0
